backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-admin-logs==1.0.2
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
